Updated event handler to use existing one
