+++ src/main/java/theflogat/technomancy/Technomancy.java
@@ -20,9 +20,6 @@ import theflogat.technomancy.proxies.CommonProxy;
 import theflogat.technomancy.util.Loc;
 import theflogat.technomancy.util.Ore;
 import net.minecraft.creativetab.CreativeTabs;
-import net.minecraft.item.Item;
-import net.minecraftforge.client.event.TextureStitchEvent;
-import net.minecraftforge.common.MinecraftForge;
 import cpw.mods.fml.common.Mod;
 import cpw.mods.fml.common.Mod.EventHandler;
 import cpw.mods.fml.common.Mod.Instance;
@@ -31,10 +28,7 @@ import cpw.mods.fml.common.event.FMLInitializationEvent;
 import cpw.mods.fml.common.event.FMLInterModComms;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
-import cpw.mods.fml.common.eventhandler.SubscribeEvent;
 import cpw.mods.fml.common.network.NetworkRegistry;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 @Mod(modid = Ref.MOD_ID, 
 	name = Ref.MOD_NAME, 
@@ -60,8 +54,6 @@ public class Technomancy {
     public void preInit(FMLPreInitializationEvent event) {
     	ConfigHandler.init(new File(event.getModConfigurationDirectory(), Ref.MOD_NAME + ".cfg"));
     	new EventRegister();
-    	
-    	MinecraftForge.EVENT_BUS.register(this);
     }
 
     @EventHandler
@@ -122,13 +114,4 @@ public class Technomancy {
         	CompatibilityHandler.smeltify();
         }
     }
-    
-    @SubscribeEvent
-	@SideOnly(Side.CLIENT)
-	public void stitchEventPost(TextureStitchEvent.Post event) {
-		if (event.map.getTextureType() == 1) {
-			Ore.initColors();
-			ConfigHandler.initColorConfigs();
-		}
-	}
 }
